// src/abi/types/IOrderBook.ts
export declare namespace Funding {
  export type GrowthStruct = {
    twPremiumX96: BigNumberish;
    twPremiumDivBySqrtPriceX96: BigNumberish;
  };

  export type GrowthStructOutput = [BigNumber, BigNumber] & {
    twPremiumX96: BigNumber;
    twPremiumDivBySqrtPriceX96: BigNumber;
  };
}

export declare namespace IOrderBook {
  export type AddLiquidityParamsStruct = {
    trader: string;
    baseToken: string;
    base: BigNumberish;
    quote: BigNumberish;
    lowerTick: BigNumberish;
    upperTick: BigNumberish;
    fundingGrowthGlobal: Funding.GrowthStruct;
  };

  export type AddLiquidityParamsStructOutput = [
    string,
    string,
    BigNumber,
    BigNumber,
    number,
    number,
    Funding.GrowthStructOutput
  ] & {
    trader: string;
    baseToken: string;
    base: BigNumber;
    quote: BigNumber;
    lowerTick: number;
    upperTick: number;
    fundingGrowthGlobal: Funding.GrowthStructOutput;
  };

  export type AddLiquidityResponseStruct = {
    base: BigNumberish;
    quote: BigNumberish;
    fee: BigNumberish;
    liquidity: BigNumberish;
  };

  export type AddLiquidityResponseStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    base: BigNumber;
    quote: BigNumber;
    fee: BigNumber;
    liquidity: BigNumber;
  };

  export type RemoveLiquidityParamsStruct = {
    maker: string;
    baseToken: string;
    lowerTick: BigNumberish;
    upperTick: BigNumberish;
    liquidity: BigNumberish;
  };

  export type RemoveLiquidityParamsStructOutput = [
    string,
    string,
    number,
    number,
    BigNumber
  ] & {
    maker: string;
    baseToken: string;
    lowerTick: number;
    upperTick: number;
    liquidity: BigNumber;
  };

  export type RemoveLiquidityResponseStruct = {
    base: BigNumberish;
    quote: BigNumberish;
    fee: BigNumberish;
    takerBase: BigNumberish;
    takerQuote: BigNumberish;
  };

  export type RemoveLiquidityResponseStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    base: BigNumber;
    quote: BigNumber;
    fee: BigNumber;
    takerBase: BigNumber;
    takerQuote: BigNumber;
  };

  export type ReplaySwapParamsStruct = {
    baseToken: string;
    isBaseToQuote: boolean;
    shouldUpdateState: boolean;
    amount: BigNumberish;
    sqrtPriceLimitX96: BigNumberish;
    exchangeFeeRatio: BigNumberish;
    uniswapFeeRatio: BigNumberish;
    globalFundingGrowth: Funding.GrowthStruct;
  };

  export type ReplaySwapParamsStructOutput = [
    string,
    boolean,
    boolean,
    BigNumber,
    BigNumber,
    number,
    number,
    Funding.GrowthStructOutput
  ] & {
    baseToken: string;
    isBaseToQuote: boolean;
    shouldUpdateState: boolean;
    amount: BigNumber;
    sqrtPriceLimitX96: BigNumber;
    exchangeFeeRatio: number;
    uniswapFeeRatio: number;
    globalFundingGrowth: Funding.GrowthStructOutput;
  };

  export type ReplaySwapResponseStruct = {
    tick: BigNumberish;
    fee: BigNumberish;
    insuranceFundFee: BigNumberish;
  };

  export type ReplaySwapResponseStructOutput = [
    number,
    BigNumber,
    BigNumber
  ] & { tick: number; fee: BigNumber; insuranceFundFee: BigNumber };
}